core(simd): add (q)narrow instructions - narrow32, narrow16, narrow8, narrow4, qnarrow32, qnarrow32u, qnarrow16, qnarrow16u, qnarrow8, qnarrow8u, qnarrow4, qnarrow4u

+++ src/core_exec_custom_simd_data_fmt_widen.cpp
@@ -2,7 +2,7 @@
 #include "core.h"
 #include "core_exec_custom_simd.h"
 
-template <int vbits, bool vsigned>
+template <size_t vbits, bool vsigned>
 reg_pair core::data_fmt_c_widen_t(uint32_t a) {
     // widen n-bit to 2n-bit elements
     constexpr size_t e = (32 / vbits);
